requirements.txt
pandas
kafka-python
requests
faker